% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceelix.R
\name{niceelix}
\alias{niceelix}
\title{Lisa's Elixhauser comobidity score calculator}
\usage{
niceelix(df, score = "mortality", imputeNA = FALSE, aids = "cm_aids",
  alcohol = "cm_alcohol", anemdef = "cm_anemdef", arth = "cm_arth",
  bldloss = "cm_bldloss", chf = "cm_chf", chrnlung = "cm_chrnlung",
  coag = "cm_coag", depress = "cm_depress", dm = "cm_dm",
  dmcx = "cm_dmcx", drug = "cm_drug", htn_c = "cm_htn_c",
  hypothy = "cm_hypothy", liver = "cm_liver", lymph = "cm_lymph",
  lytes = "cm_lytes", mets = "cm_mets", neuro = "cm_neuro",
  obese = "cm_obese", para = "cm_para", perivasc = "cm_perivasc",
  psych = "cm_psych", pulmcirc = "cm_pulmcirc",
  renlfail = "cm_renlfail", tumor = "cm_tumor", ulcer = "cm_ulcer",
  valve = "cm_valve", wghtloss = "cm_wghtloss")
}
\arguments{
\item{df}{Dataframe [REQUIRED].}

\item{score}{Character. Name of Elixhauser score to calculate ("mortality" or "readmission"). Default is "mortality".}

\item{imputeNA}{Logical. Should missing values be imputed as 0? Default is FALSE.}

\item{aids}{Character. Name of comorbidity indicator variable in dataframe.}

\item{alcohol}{Character. Name of comorbidity indicator variable in dataframe.}

\item{anemdef}{Character. Name of comorbidity indicator variable in dataframe.}

\item{arth}{Character. Name of comorbidity indicator variable in dataframe.}

\item{bldloss}{Character. Name of comorbidity indicator variable in dataframe.}

\item{chf}{Character. Name of comorbidity indicator variable in dataframe.}

\item{chrnlung}{Character. Name of comorbidity indicator variable in dataframe.}

\item{coag}{Character. Name of comorbidity indicator variable in dataframe.}

\item{depress}{Character. Name of comorbidity indicator variable in dataframe.}

\item{dm}{Character. Name of comorbidity indicator variable in dataframe.}

\item{dmcx}{Character. Name of comorbidity indicator variable in dataframe.}

\item{drug}{Character. Name of comorbidity indicator variable in dataframe.}

\item{htn_c}{Character. Name of comorbidity indicator variable in dataframe.}

\item{hypothy}{Character. Name of comorbidity indicator variable in dataframe.}

\item{liver}{Character. Name of comorbidity indicator variable in dataframe.}

\item{lymph}{Character. Name of comorbidity indicator variable in dataframe.}

\item{lytes}{Character. Name of comorbidity indicator variable in dataframe.}

\item{mets}{Character. Name of comorbidity indicator variable in dataframe.}

\item{neuro}{Character. Name of comorbidity indicator variable in dataframe.}

\item{obese}{Character. Name of comorbidity indicator variable in dataframe.}

\item{para}{Character. Name of comorbidity indicator variable in dataframe.}

\item{perivasc}{Character. Name of comorbidity indicator variable in dataframe.}

\item{psych}{Character. Name of comorbidity indicator variable in dataframe.}

\item{pulmcirc}{Character. Name of comorbidity indicator variable in dataframe.}

\item{renlfail}{Character. Name of comorbidity indicator variable in dataframe.}

\item{tumor}{Character. Name of comorbidity indicator variable in dataframe.}

\item{ulcer}{Character. Name of comorbidity indicator variable in dataframe.}

\item{valve}{Character. Name of comorbidity indicator variable in dataframe.}

\item{wghtloss}{Character. Name of comorbidity indicator variable in dataframe.}
}
\description{
This function calculates the mortality or readmission Elixhauser comorbidity score from individual comorbidity indicator variables.
Input a data frame and the variable names for each of the comorbidity indicator variables.
This function outputs a numeric vector of scores.
}
\keyword{Elixhauser}
\keyword{comorbidity}
\keyword{index}
\keyword{score}
